Adds the ability to override the level in the IslandLevelCalculatedEvent
https://github.com/BentoBoxWorld/Level/issues/125

+++ src/main/java/world/bentobox/level/calculators/CalcIslandLevel.java
@@ -390,7 +390,7 @@ public class CalcIslandLevel {
          * Set level
          * @param level - level
          */
-        public void setLevel(int level) {
+        public void setLevel(long level) {
             this.level.set(level);
         }
         /**
